New configuration with externalized yaml files for every model
Now we lazily load model configurations from kge/model and also lazily
import the models themselves. So no more need to define the models also
in __init__ and no more if else in kge_model necessary. The only files
that are now relevant for a new model is MODEL_NAME.py and
MODEL_NAME.yaml. To resolve the class name the model config now has a
mandatory field class_name that will be used for the import.

Additionally +++ is now +++: +++ in the core schema config, to allow for
some preconfiguration. We should not put a default learning rate there
as Pytorch already has a default learning rate for every optimizer and
every optimizer has different learning rates, so it could even lead to
unwanted behaviour that a user leaves the lr for Adam blank and gets the
default learning rate from the config which would likely be way too
large. E.g. Adagrad (0.1-0.01) Adam (0.001-0.0001).