Revised handling of output folders
Output folders and filenames are not part of the configuration anymore; this was
asking for trouble.

Folders are not taken relative the the kge directory anymore, but relative to
the current directory. This improves usability: one can now run kge.py from
whereever desired. E.g., in a directory that contains ```toy.yaml```:

```
<kge-dir>/kge.py create toy.yaml -f ~/toy-result
```

and later in ```~/toy-result```

```
<kge-dir>/kge.py resume .
```

or from anywhere

```
<kge-dir>/kge.py resume ~/toy.result
```